renamed ead space to reflect more generally what the stuff in this namespace is doing
