+++ fda-model-test/simulationShould.cs
@@ -1,7 +1,7 @@
 ﻿
 
 using Xunit;
-using ead;
+using compute;
 using paireddata;
 using Statistics;
 using System.Collections.Generic;
@@ -50,7 +50,7 @@ namespace fda_model_test
                 .withStageDamages(upd)
                 .withAdditionalThreshold(threshold)
                 .build();
-            ead.MeanRandomProvider mrp = new MeanRandomProvider();
+            compute.MeanRandomProvider mrp = new MeanRandomProvider();
             metrics.Results r = s.Compute(mrp,1);
             double difference = expected - r.ExpectedAnnualDamageResults.MeanEAD("residential");
             double relativeDifference = difference / expected;
@@ -130,7 +130,7 @@ namespace fda_model_test
                 .withStageDamages(upd)
                 .build();
             //Simulation s = new Simulation(flow_frequency, flow_stage, levee, upd);
-            ead.MeanRandomProvider mrp = new MeanRandomProvider();
+            compute.MeanRandomProvider mrp = new MeanRandomProvider();
             metrics.Results r = s.Compute(mrp, 1);
             double difference = expected - r.ExpectedAnnualDamageResults.MeanEAD("residential");
             double relativeDifference = difference / expected;
