+++ fda-model/compute/EADMessage.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace ead
+namespace compute
 {
     public class EADMessage : Base.Interfaces.IMessage
     {
